Cleaned up codes
