+++ src/main/java/me/zeroeightsix/kami/util/graphics/font/CharInfo.kt
@@ -1,0 +1,21 @@
+package me.zeroeightsix.kami.util.graphics.font
+
+class CharInfo(
+    /** Character's width  */
+    val width: Double,
+
+    /** Character's height  */
+    val height: Double,
+
+    /** Upper left u */
+    val u1: Double,
+
+    /** Upper left v */
+    val v1: Double,
+
+    /** Lower right u */
+    val u2: Double,
+
+    /** Lower right v */
+    val v2: Double
+)
